pyproject.toml
[project]
name = "stock-data-collector"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.26.4",
    "pandas-stubs==2.3.0.250703",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tushare>=1.4.23",
]
